+++ test/compile/index.js
@@ -2,7 +2,7 @@ const assert = require('assert');
 const compile = require('../../src/compile/index');
 const tplTokenizer = require('../../src/compile/tpl-tokenizer');
 const defaults = require('../../src/compile/defaults');
-const debug = defaults.debug;
+const debuger = defaults.debuger;
 const path = require('path');
 
 let render, data, result;
@@ -11,13 +11,13 @@ module.exports = {
 
     before: () => {
         console.log('#compile/index');
-        defaults.debug = () => {
+        defaults.debuger = () => {
             return () => '{Template Error}';
         };
     },
 
     after: () => {
-        defaults.debug = debug;
+        defaults.debuger = debuger;
     },
 
     'rule.native': {