+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 1. 调试功能增强：支持编译阶段捕获语法错误具体行
 2. `template.config()` 方法取消，请直接读写 `template.defaults`
 3. `template.helpers` 更改为 `template.defaults.imports`
-4. `template.defaults.debug` 选项可以定制输出的调试信息
+4. `template.defaults.debuger` 选项可以定制输出的调试信息
 5. `template.defaults.parser` 被取消，使用更强大的 `template.defaults.rules` 代替
 6. 兼容 EJS 的模板
 7. 使用 ECMA5 构建，低于 IE9 的浏览器需要使用 ECMA5 补丁以及 JSON 库支持才可以运行